refactoring calibration code
