+++ src/mag_toolkit/calibration/CalibrationDefinitions.py
@@ -1,0 +1,52 @@
+from abc import ABC
+from datetime import datetime
+from enum import Enum
+from typing import Optional
+
+from pydantic import BaseModel
+
+
+class CalibrationMethod(str, Enum):
+    KEPKO = "Kepko"
+    LEINWEBER = "Leinweber"
+    IMAPLO_PIVOT = "IMAP-Lo Pivot Platform Interference"
+    NOOP = "noop"
+    SUM = "Sum of other calibrations"
+
+
+class Sensor(str, Enum):
+    MAGO = "MAGo"
+    MAGI = "MAGi"
+
+
+class Mission(str, Enum):
+    IMAP = "IMAP"
+
+
+class CalibrationMetadata(BaseModel):
+    dependencies: list[str]
+    science: list[str]
+    creation_timestamp: datetime
+    comment: Optional[str] = None
+
+
+class Value(BaseModel, ABC):
+    time: datetime
+    value: list[float]
+
+
+class CalibrationValue(Value):
+    timedelta: float = 0
+    quality_flag: int = 0
+    quality_bitmask: int = 0
+
+
+class ScienceValue(Value):
+    range: int
+    quality_flag: Optional[int] = 0
+    quality_bitmask: Optional[int] = 0
+
+
+class Validity(BaseModel):
+    start: datetime
+    end: datetime
