+++ src/mag_toolkit/calibration/calibrators/__init__.py
@@ -1,0 +1,7 @@
+from .Calibrator import Calibrator
+from .EmptyCalibrator import EmptyCalibrator
+
+__all__ = [
+    "Calibrator",
+    "EmptyCalibrator",
+]
